requirements.txt
requests
python-dotenv
openai
pyyaml
pdfplumber
flask
gunicorn